fix(flow): avoid signed size comparison in store

+++ Telegram/SourceFiles/teleflow/flow/flow_store.cpp
@@ -121,7 +121,7 @@ AddResult AddItem(
 			return AddResult::AlreadyExists;
 		}
 	}
-	if (items.size() >= kMaxItems) {
+	if (items.size() >= size_t(kMaxItems)) {
 		return AddResult::StorageLimitReached;
 	}
